third_party/blink/web_tests/fast/xsl/default-html-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
Test for bug 11694: XSLT output method does not default to HTML when the target document is HTML.

Should be green: result.
Should be black: result.

SUCCESS
